dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=dev-requirements.txt pyproject.toml
#
asgiref==3.8.1
    # via django
attrs==24.2.0
    # via pytest
bleach[css]==5.0.1
    # via
    #   chipy_website (pyproject.toml)
    #   django-bleach
boto3==1.35.62
    # via chipy_website (pyproject.toml)
botocore==1.35.62
    # via
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.0
    # via requests
crispy-bootstrap5==2024.10
    # via chipy_website (pyproject.toml)
cryptography==43.0.3
    # via
    #   python-openid2
    #   social-auth-core
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
distlib==0.3.9
    # via virtualenv
dj-database-url==2.3.0
    # via chipy_website (pyproject.toml)
django==5.1.3
    # via
    #   chipy_website (pyproject.toml)
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-bleach
    #   django-crispy-forms
    #   django-extensions
    #   django-honeypot
    #   django-ical
    #   django-recaptcha
    #   django-recurrence
    #   django-storages
    #   django-tinymce
    #   djangorestframework
    #   social-auth-app-django
django-admin-tools==0.9.3
    # via chipy_website (pyproject.toml)
django-bleach==3.1.0
    # via chipy_website (pyproject.toml)
django-crispy-forms==2.3
    # via
    #   chipy_website (pyproject.toml)
    #   crispy-bootstrap5
django-extensions==3.2.3
    # via chipy_website (pyproject.toml)
django-flatblocks==1.0.0
    # via chipy_website (pyproject.toml)
django-grappelli==4.0.1
    # via chipy_website (pyproject.toml)
django-gravatar2==1.4.5
    # via chipy_website (pyproject.toml)
django-honeypot==1.2.1
    # via chipy_website (pyproject.toml)
django-ical==1.9.2
    # via chipy_website (pyproject.toml)
django-recaptcha==4.0.0
    # via chipy_website (pyproject.toml)
django-recurrence==1.11.1
    # via django-ical
django-storages==1.14.4
    # via chipy_website (pyproject.toml)
django-tinymce==4.1.0
    # via chipy_website (pyproject.toml)
djangorestframework==3.15.2
    # via chipy_website (pyproject.toml)
doublemetaphone==1.1
    # via probablepeople
filelock==3.16.1
    # via virtualenv
freezegun==1.5.1
    # via chipy_website (pyproject.toml)
gunicorn==23.0.0
    # via chipy_website (pyproject.toml)
icalendar==6.0.1
    # via django-ical
identify==2.6.3
    # via pre-commit
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.3.0
    # via python-openid2
nodeenv==1.9.1
    # via pre-commit
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.2
    # via
    #   gunicorn
    #   pytest
pillow==11.0.0
    # via chipy_website (pyproject.toml)
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via chipy_website (pyproject.toml)
probableparsing==0.0.1
    # via probablepeople
probablepeople==0.5.6
    # via chipy_website (pyproject.toml)
psycopg2-binary==2.9.10
    # via chipy_website (pyproject.toml)
py==1.11.0
    # via pytest
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via social-auth-core
pytest==6.2.5
    # via
    #   chipy_website (pyproject.toml)
    #   pytest-django
pytest-django==4.5.2
    # via chipy_website (pyproject.toml)
python-crfsuite==0.9.11
    # via probablepeople
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   django-recurrence
    #   freezegun
    #   icalendar
python-openid2==3.2
    # via chipy_website (pyproject.toml)
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.2
    # via pre-commit
requests==2.32.3
    # via
    #   chipy_website (pyproject.toml)
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via
    #   bleach
    #   chipy_website (pyproject.toml)
    #   python-dateutil
    #   python-openid2
social-auth-app-django==5.4.2
    # via chipy_website (pyproject.toml)
social-auth-core==4.5.4
    # via social-auth-app-django
sorl-thumbnail==12.11.0
    # via chipy_website (pyproject.toml)
sqlparse==0.5.2
    # via django
tinycss2==1.1.1
    # via bleach
toml==0.10.2
    # via pytest
typing-extensions==4.12.2
    # via dj-database-url
tzdata==2024.2
    # via icalendar
urllib3==2.2.3
    # via
    #   botocore
    #   requests
uwsgi==2.0.28
    # via chipy_website (pyproject.toml)
virtualenv==20.28.0
    # via pre-commit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
whitenoise==6.8.2
    # via chipy_website (pyproject.toml)
